Revert "Home Work 2.12  News API"
This reverts commit e6a7599b357f557d0b18ac2b5a56b02df0f7ea61.
