--- NewsAPI/NewsAPI.xcodeproj/project.xcworkspace/xcuserdata/katerina.xcuserdatad/IDEFindNavigatorScopes.plist
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<array/>
-</plist>
