+++ NewsAPI/NewsAPI/AppDelegate.swift
@@ -15,8 +15,8 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
     
         self.window = UIWindow(frame: UIScreen.main.bounds)
-       
-        if let window = window {   // разворачиваем
+        // разворачиваем
+        if let window = window {
             let tabBar = TabBarController()
             window.rootViewController = tabBar
             window.makeKeyAndVisible()
