+++ NewsAPI/NewsAPI/GeneralVC/View/GeneralCollectionViewCell.swift
