+++ NewsAPI/NewsAPI/Model/ArticleResponceObject.swift
@@ -9,8 +9,8 @@ import Foundation
 
 struct ArticleResponceObject: Codable {
     let title: String
-    let description: String?
-    let urlToImage: String?
+    let description: String
+    let urlToImage: String
     let date: String
     
     enum CodingKeys: String, CodingKey {
