+++ NewsAPI/NewsAPI/TabBarController.swift
@@ -20,20 +20,17 @@ final class TabBarController: UITabBarController {
     private func setupViewControllers() {
         
         viewControllers = [
-            
             setupNavigationController(rootViewController: GeneralViewController(viewModel: GeneralViewModel()),
                                       title: "General",
                                       image: UIImage(systemName: "newspaper") ?? UIImage.add),
-            setupNavigationController(rootViewController: BusinessViewController(viewModel: BusinessViewModel()),
+            setupNavigationController(rootViewController: BusinessViewController(),
                                       title: "Business",
                                       image: UIImage(systemName: "briefcase") ?? UIImage.add),
-            
-            setupNavigationController(rootViewController: TechnologyViewController(viewModel: TechnologyViewModel()),
-                                      title: "Technology",
+            setupNavigationController(rootViewController: TechnologyViewController(),
+                                      title: "Technologi",
                                       image: UIImage(systemName: "gyroscope") ?? UIImage.add)
         ]
     }
-    
     private func setupNavigationController(rootViewController: UIViewController,
                                            title: String,
                                            image: UIImage) -> UINavigationController {
@@ -46,8 +43,8 @@ final class TabBarController: UITabBarController {
         navigationController.navigationBar.prefersLargeTitles = true
         
         return navigationController
+        
     }
-    
     private func setupTabBar() {
         let appearance = UITabBarAppearance()
         appearance.configureWithOpaqueBackground()
