+++ NewsAPI/NewsAPI/TechnologiVC/TechnologyViewController.swift
@@ -1,0 +1,27 @@
+//
+//  TechnologyViewController.swift
+//  NewsAPI
+//
+//  Created by Katerina on 05/12/2023.
+//
+
+import UIKit
+
+final class TechnologyViewController: UIViewController {
+    
+    
+     // MARK: - GUI Variables,  // добавляем во viewDidLoad() и настраиваем констрейнты
+     
+     // MARK: - Properties
+     
+     // MARK: - Life circle
+     
+     override func viewDidLoad() {
+         super.viewDidLoad()
+         view.backgroundColor = .blue
+     }
+     
+     // MARK: - Methods
+
+     // MARK: - Private methods
+}
