--- NewsAPI/NewsAPI/VC/BusinessVC/View/BusinessViewController.swift
@@ -1,151 +1,0 @@
-//
-//  BusinessViewController.swift
-//  NewsAPI
-//
-//  Created by Katerina on 05/12/2023.
-//
-
-import UIKit
-import SnapKit
-
-class BusinessViewController: BaseNewsVC {
-    
-    // MARK: - GUI Variables
-    private lazy var collectionView: UICollectionView = {
-        let layout = UICollectionViewFlowLayout()
-     
-        layout.minimumLineSpacing = 5
-        layout.minimumInteritemSpacing = 5
-        layout.sectionInset = UIEdgeInsets(top: 20,
-                                           left: 20,
-                                           bottom: 20,
-                                           right: 20)
-        
-        let collectionView = UICollectionView(frame: CGRect(x: 0,
-                                                            y: 0,
-                                                            width: view.frame.width,
-                                                            height: view.frame.height),
-                                              collectionViewLayout: layout)
-        collectionView.dataSource = self
-        collectionView.delegate = self
-        collectionView.backgroundColor = .white
-        
-        return collectionView
-    }()
-    
-    // MARK: - Properties
-    private var viewModel: NewsListViewModelProtocol
-    
-    // MARK: - Life circle
-    init(viewModel: NewsListViewModelProtocol) {
-        self.viewModel = viewModel
-        super.init(nibName: nil, bundle: nil)
-        self.setupViewModel()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        setupUI()
-        registerCell(collectionView: collectionView)
-        viewModel.loadData(searchText: nil)
-    }
-    
-    // MARK: - Methods
-    private func setupViewModel() {
-        viewModel.reloadData = { [weak self] in
-            self?.collectionView.reloadData()
-        }
-        
-        viewModel.reloadCell = { [weak self] indexPath in
-            self?.collectionView.reloadItems(at: [indexPath])
-        }
-        
-        viewModel.showError = { error in
-            self.showAlert(message: error)
-            print(error)
-        }
-    }
-    
-    // MARK: - Private methods
-    private func setupUI() {
-        setupStyle()
-        view.addSubview(collectionView)
-        
-        setupConstraints()
-    }
-    
-    private func setupConstraints() {
-        collectionView.snp.makeConstraints { make in
-            
-            make.leading.trailing.equalToSuperview().inset(5)
-            make.top.bottom.equalTo(view.safeAreaLayoutGuide)
-        }
-    }
-}
-
-// MARK: - UICollectionViewDataSource - методы таблицы
-extension BusinessViewController: UICollectionViewDataSource {
-    func numberOfSections(in collectionView: UICollectionView) -> Int {
-        viewModel.sections.count
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        viewModel.sections[section].items.count
-    }
-    //отображение реальной инфы: set новостей в эти ячейки
-    func collectionView(_ collectionView: UICollectionView,
-                        cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        guard let article = viewModel.sections[indexPath.section].items[indexPath.row] as? ArticleCellViewModel else { return UICollectionViewCell() }
-        
-        if indexPath.section == 0 {
-            let cell =  collectionView.dequeueReusableCell(withReuseIdentifier: "GeneralCollectionViewCell", for: indexPath) as? GeneralCollectionViewCell
-            cell?.set(article: article)
-            
-            return cell ?? UICollectionViewCell()
-        } else {
-            let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "DetailsCollectionViewCell", for: indexPath) as? DetailsCollectionViewCell
-            cell?.set(article: article)
-            
-            return cell ?? UICollectionViewCell()
-        }
-    }
-}
-
-// MARK: - UICollectionViewDelegate
-extension BusinessViewController: UICollectionViewDelegate {
-    func collectionView(_ collectionView: UICollectionView,
-                        didSelectItemAt indexPath: IndexPath) {
-        guard let article = viewModel.sections[indexPath.section].items[indexPath.row] as? ArticleCellViewModel else { return  }
-        
-        navigationController?.pushViewController(InfoVC(viewModel: NewsViewModel(article: article)), animated: true)
-    }
-    
-    func collectionView(_ collectionView: UICollectionView,
-                        willDisplay cell: UICollectionViewCell,
-                        forItemAt indexPath: IndexPath) {
-        if indexPath.row == (viewModel.sections[indexPath.section].items.count - 15) {
-            viewModel.loadData(searchText: nil)
-        }
-    }
-}
-
-// MARK: - UICollectionViewDelegateFlowLayout
-extension BusinessViewController: UICollectionViewDelegateFlowLayout {
-    func collectionView(_ collectionView: UICollectionView,
-                        layout collectionViewLayout: UICollectionViewLayout,
-                        sizeForItemAt indexPath: IndexPath) -> CGSize {
-        let width = view.frame.width
-        let firstSectionItemSize = CGSize(width: width, height: width)
-        let secondSectionItemSize = CGSize(width: width, height: 100)
-        
-        return indexPath.section == 0 ? firstSectionItemSize : secondSectionItemSize
-    }
-}
-
-
-
