--- NewsAPI/NewsAPI/VC/GeneralVC/ViewModel/GeneralViewModel.swift
@@ -1,20 +1,0 @@
-//
-//  GeneralViewModel.swift
-//  NewsAPI
-//
-//  Created by Katerina on 11/12/2023.
-//
-
-import Foundation
-
-final class GeneralViewModel: NewsListViewModel {
-    override func loadData(searchText: String?) {
-        super.loadData(searchText: searchText)
-        
-        ApiManager.getNews(from: .general,
-                           page: page,
-                           searchText: searchText) { [weak self] result in
-            self?.handleResult(result)
-        }
-    }
-}
