--- NewsAPI/NewsAPI/VC/TechnologiVC/TechnologyViewController.swift
@@ -1,125 +1,0 @@
-//
-//  TechnologyViewController.swift
-//  NewsAPI
-//
-//  Created by Katerina on 05/12/2023.
-//
-
-import UIKit
-
-final class TechnologyViewController: BaseNewsVC {
-    
-    // MARK: - GUI Variables
-    private lazy var collectionView: UICollectionView = {
-        let layout = UICollectionViewFlowLayout()
-        
-        let width = (view.frame.width - 10)
-        layout.itemSize = CGSize(width: width, height: width/3)
-        
-        layout.minimumLineSpacing = 5
-        layout.minimumInteritemSpacing = 5
-        
-        let collectionView = UICollectionView(frame: CGRect(x: 0,
-                                                            y: 0,
-                                                            width: view.frame.width,
-                                                            height: view.frame.height),
-                                              collectionViewLayout: layout)
-        collectionView.dataSource = self
-        collectionView.delegate = self
-        collectionView.backgroundColor = .clear
-        
-        return collectionView
-    }()
-    
-    // MARK: - Properties
-    private var viewModel: NewsListViewModelProtocol
-    
-    init(viewModel: NewsListViewModelProtocol) {
-        self.viewModel = viewModel
-        super.init(nibName: nil, bundle: nil)
-        self.setupViewModel()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    // MARK: - Life circle
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        setupUI()
-        registerCell()
-        viewModel.loadData(searchText: nil)
-    }
-    
-    // MARK: - Methods
-    func registerCell() {
-        collectionView.register(TechnologyCollectionViewCell.self, forCellWithReuseIdentifier: "TechnologyCollectionViewCell")
-    }
-    
-    private func setupViewModel() {
-        viewModel.reloadData = { [weak self] in
-            self?.collectionView.reloadData()
-        }
-        
-        viewModel.reloadCell = { [weak self] indexPath in
-            self?.collectionView.reloadItems(at: [indexPath])
-        }
-        
-        viewModel.showError = { error in
-            self.showAlert(message: error)
-            print(error)
-        }
-    }
-    
-    // MARK: - Private methods
-    private func setupUI() {
-        setupStyle()
-        
-        view.addSubview(collectionView)
-        setupConstraints()
-    }
-    
-    private func setupConstraints() {
-        collectionView.snp.makeConstraints { make in
-            make.top.equalToSuperview()
-            make.leading.trailing.equalToSuperview().inset(5)
-            make.bottom.equalTo(view.safeAreaLayoutGuide)
-        }
-    }
-}
-
-// MARK: - UICollectionViewDataSource
-extension TechnologyViewController: UICollectionViewDataSource {
-    func numberOfSections(in collectionView: UICollectionView) -> Int {
-        viewModel.sections.count
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        viewModel.sections[section].items.count
-    }
-    
-    func collectionView(_ collectionView: UICollectionView,
-                        cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        
-        guard let article = viewModel.sections[indexPath.section].items[indexPath.row] as? ArticleCellViewModel,
-              let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "TechnologyCollectionViewCell",
-                                                            for: indexPath) as? TechnologyCollectionViewCell else
-        { return UICollectionViewCell() }
-        cell.set(article: article)
-        
-        return cell
-    }
-}
-
-// MARK: - UICollectionViewDelegate
-extension TechnologyViewController: UICollectionViewDelegate {
-    func collectionView(_ collectionView: UICollectionView,
-                        didSelectItemAt indexPath: IndexPath) {
-        guard let article = viewModel.sections[indexPath.section].items[indexPath.row] as? ArticleCellViewModel else { return  }
-        
-        navigationController?.pushViewController(InfoVC(viewModel: NewsViewModel(article: article)),
-                                                 animated: true)
-    }
-}
